Added users table and authentification feature with token

+++ Python/Blogging-Platform-API/src/routers/post.py
@@ -1,5 +1,5 @@
 from fastapi import APIRouter, status, Depends, HTTPException, Response
-from .. import schemas, models
+from .. import schemas, models, oauth2
 from sqlalchemy.orm import Session
 from ..database import get_db
 from typing import List, Optional
@@ -11,9 +11,9 @@ router = APIRouter(
 )
 
 @router.post("/", status_code=status.HTTP_201_CREATED, response_model=schemas.Post)
-def create_post(post: schemas.PostCreate, db: Session = Depends(get_db)):
+def create_post(post: schemas.PostCreate, current_user: int=Depends(oauth2.get_current_user), db: Session = Depends(get_db)):
     
-    new_post = models.Post(**post.dict())
+    new_post = models.Post(owner_id=current_user.id, **post.model_dump())
     
     db.add(new_post)
     db.commit()
@@ -22,7 +22,7 @@ def create_post(post: schemas.PostCreate, db: Session = Depends(get_db)):
     return new_post
 
 @router.put("/{id}", response_model=schemas.Post)
-def update_post(id: int, updated_post: schemas.PostCreate, db: Session = Depends(get_db)):
+def update_post(id: int, updated_post: schemas.PostCreate, current_user: int=Depends(oauth2.get_current_user), db: Session = Depends(get_db)):
     
     post_query = db.query(models.Post).filter(models.Post.id == id)
     post = post_query.first()
@@ -30,19 +30,25 @@ def update_post(id: int, updated_post: schemas.PostCreate, db: Session = Depends
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Post with id {id} does not exist")
     
-    post_query.update(updated_post.dict(), synchronize_session=False)
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=f"Not authorized to perform requested action")
+    
+    post_query.update(updated_post.model_dump(), synchronize_session=False)
     db.commit()
     
     return post_query.first()
 
 @router.delete("/{id}", status_code=status.HTTP_204_NO_CONTENT)
-def delete_post(id: int, db: Session = Depends(get_db)):
+def delete_post(id: int, current_user: int=Depends(oauth2.get_current_user), db: Session = Depends(get_db)):
     
     post_query = db.query(models.Post).filter(models.Post.id == id)
     post = post_query.first()
     
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Post with id {id} does not exist")
+    
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=f"Not authorized to perform requested action")
     
     post_query.delete(synchronize_session=False)
     db.commit()
@@ -52,6 +58,7 @@ def delete_post(id: int, db: Session = Depends(get_db)):
 @router.get("/", response_model=List[schemas.Post])
 def get_posts(db: Session = Depends(get_db), limit: int=10, skip: int=0, term: Optional[str]=""):
     
+    # Search across title, content and category with optional term, default returns all
     posts = db.query(models.Post).filter(or_(models.Post.title.contains(term), models.Post.content.contains(term), models.Post.category.contains(term))).limit(limit).offset(skip).all()
     
     return posts
